autherization & authentication checked with JWT using postman api

+++ controllers/api/v1/post_api.js
@@ -25,7 +25,7 @@ module.exports.destroy =async(req,res)=>{
           const post =await Post.findById(req.params.id);//from url -/posts/destroy/:id
           // req.user._id converted to string my mongoose req.user.id
           // also post.user is ref to user i.e userid
-   
+         if(post.user == req.user.id){
             await post.deleteOne();            // delete all comment associated with this post
            const isCommentDltd= await Comment.deleteMany({post: req.params.id});
           
@@ -34,7 +34,11 @@ module.exports.destroy =async(req,res)=>{
             message:"post and its associated comments got deleted",
             
         });            
-           
+    }else{
+        return res.json(401,{
+            message:"You cannot delete this post"
+        })
+    }   
 
           
     }catch(err){
